+++ flake.nix
@@ -252,6 +252,11 @@
           cargoBuildFlags = [ "--package" "d2b-clipd" "--bin" "d2b-clipd" ];
           doCheck = false;
         };
+        d2b-wayland-proxy = rustWorkspace {
+          pname = "d2b-wayland-proxy";
+          cargoBuildFlags = [ "--package" "d2b-wayland-proxy" "--bin" "d2b-wayland-proxy" ];
+          doCheck = false;
+        };
 
         signoz = import ./pkgs/signoz { inherit pkgs; };
         signozOtelCollector = import ./pkgs/signoz-otel-collector { inherit pkgs; };